premier

+++ Bilel/old/premier.c
@@ -3,16 +3,16 @@
 
 int main (int argc, char *argv[])
 {
-    int n = 0;
+    long n = 0;
     int r = 0;
 
   printf("***** Nombre Premier ****** \n");
   printf(" Choose a Number : ");
-  scanf("%d",&n);
+  scanf("%ld",&n);
 
   if (n <= 1)
   {
-    printf("%d, n'est pas premier \n",n);
+    printf("%ld, n'est pas premier \n",n);
     return 0;
   }
   
@@ -25,10 +25,10 @@ int main (int argc, char *argv[])
   }
   if (r > 2 )
   {
-    printf("%d --> n'est pas un nombre Premier.\n",n);
+    printf("%ld --> n'est pas un nombre Premier.\n",n);
     printf("nombre de Diviseur = %d\n",r);
   }else{
-    printf("%d --> Est un nombre Premier.\n",n);
+    printf("%ld --> Est un nombre Premier.\n",n);
   }
 
   return 0;
